Cache request and some code separation
